% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{fixlongnames}
\alias{fixlongnames}
\title{Fix long strongs.}
\usage{
fixlongnames(pnames, delimiter, threshold)
}
\arguments{
\item{pnames}{character vector containing strings with long names. If longer than \code{threshold} characters, will split in half.}

\item{delimiter}{string. the character separating words in the strings in pnames. default = '_'}

\item{threshold}{integer. strings of length > threshold are cut in half with a new-line at the \code{delimiter} closest to the center of the string.}
}
\description{
Long strings may contain groups of words separated by underscores or other characters, such as pathway names. This function will try to cut the string in half, replacing the delimiter character closest to the middle of the string with a new-line. Good for plottng things.
}
