% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gsea.dotplot.onecol}
\alias{gsea.dotplot.onecol}
\title{Plot GSEA results as a dotplot}
\usage{
gsea.dotplot.onecol(
  gseares,
  filter_nonsig_pathways = NULL,
  top_pathways = NULL,
  ntop = NULL,
  fix_long_pathway_names = NULL,
  pathwayfontsize = NULL,
  gsubpattern = NULL
)
}
\arguments{
\item{gseares}{dataframe with output of FGSEA. the first list element output of gsea.results.}

\item{filter_nonsig_pathways}{T/F, whether to show only signifcant pathways (FDR <0.25); default = F}

\item{top_pathways}{T/F, whether to show only top significant pathways. will show top "ntop" pathways. default = T}

\item{ntop}{integer. if top_pathways=T, how many pathways to show in each "direction" of NES. If set to 25, will show 50 total pathways (25 positive NES, 25 negative NES). default = 25.}

\item{fix_long_pathway_names}{T/F. Add newlines to pathway names if they exceed 50 characters. will try not to split words, assuming words are separated by underscores.}

\item{pathwayfontsize}{numeric. Font size for pathway names. default = 10.}

\item{gsubpattern}{character. Set this if there is a very repeptive symbol or prefix in all the pathway names. Default is "HALLMARK_"}
}
\value{
a ggplot object.
}
\description{
Visualize GSEA results with a dotplot. Dot color is given by normalized enrichment score; size is given by number of genes shared in both pathway and in dataset; and asterisks show significance.
}
