% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{modulescore}
\alias{modulescore}
\title{Calculate a sample-wise module score given a list of genes}
\usage{
modulescore(gem, genelist, numbins = NULL, numcontrolgenesperbin = NULL)
}
\arguments{
\item{gem}{matrix, a gene expresssion matrix with rows=genes and columns=samples}

\item{genelist}{character vector, genes for which to calculate module score}

\item{numbins}{integer, how many bins to split the input genelist into, for selection of control genes}

\item{numcontrolgenesperbin}{integer, how many control genes to select per bin}
}
\value{
a vector of numerics, sample-wise module scores
}
\description{
this is inspired by Tirosh et al, Science (2016) and the AddModuleScore function in Seurat
}
