% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayanalysis.R
\name{pathwayanalysis_declare_significant_clusters}
\alias{pathwayanalysis_declare_significant_clusters}
\title{Declare significant clusters in the output of pathway analysis function.}
\usage{
pathwayanalysis_declare_significant_clusters(pathwayanalysis_out, sigclusts)
}
\arguments{
\item{pathwayanalysis_out}{the output of pathwayanalysis function}

\item{sigclusts}{vector of integers. the rows in pathwayanalysis_out[\link{3}] to use as significant clusters. If not provided, will declare clusters with > 10\% of DEGs as significant.}
}
\value{
an updated pathwayanalysis_out object.
}
\description{
Set the signficant clusters in the pathwayanalysis_out object. Will return an updated version of pathwayanalysis_out[\link{3}] object with significance (T/F)
}
