% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayanalysis.R
\name{pathwayanalysis_finalize_emap_plots}
\alias{pathwayanalysis_finalize_emap_plots}
\title{Add more emapplots (graph plots with edges between similar pathways) containing cluster info.}
\usage{
pathwayanalysis_finalize_emap_plots(pathwayanalysis_out, palette)
}
\arguments{
\item{pathwayanalysis_out}{output from the \code{pathwayanalysis()} function.}

\item{palette}{character vector of colors to use (hex codes)}
}
\value{
updated \code{pathwayanalysis_out} object with plots added to \code{pathwayanalysis_out[[2]]}
}
\description{
Add more emapplots (graph plots with edges between similar pathways) containing cluster info.
}
