% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayanalysis.R
\name{pathwayanalysis_relabel_significant_clusters}
\alias{pathwayanalysis_relabel_significant_clusters}
\title{Relabel some pathway clusters with biologically relevant names}
\usage{
pathwayanalysis_relabel_significant_clusters(
  pathwayanalysis_out,
  clusterlabels,
  clusters_to_label
)
}
\arguments{
\item{pathwayanalysis_out}{output of the \code{pathwayanalysis} function.}

\item{clusterlabels}{character vector; labels to give to the clusters of pathways}

\item{clusters_to_label}{vector of integers; which clusters to re-name.}
}
\value{
will return a list, similar in format to the output of \code{pathwayanalysis} function, but with updated cluster labels in element 3.
}
\description{
Relabel some pathway clusters with biologically relevant names
}
