% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayanalysis.R
\name{pathwayanalysis_sigcluster_subplots}
\alias{pathwayanalysis_sigcluster_subplots}
\title{Make plots for pathways in each cluster.}
\usage{
pathwayanalysis_sigcluster_subplots(pathwayanalysis_out, sigres)
}
\arguments{
\item{pathwayanalysis_out}{output from the \code{pathwayanalysis()} function.}

\item{sigres}{}
}
\description{
Make plots for pathways in each cluster.
}
