% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayanalysis.R
\name{pathwayanalysis_venn}
\alias{pathwayanalysis_venn}
\title{Make a Venn Diagram with genes from the significant clusters.}
\usage{
pathwayanalysis_venn(pathwayanalysis_out)
}
\arguments{
\item{pathwayanalysis_out}{the output of the pathwayanalysis function.}
}
\description{
Please run \code{pathwayanalysis_declare_significant_clusters} first.
}
