% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayanalysis.R
\name{pathwayanalysis_wordcloud}
\alias{pathwayanalysis_wordcloud}
\title{A fast way to make a wordcloud from a character vector of pathway names}
\usage{
pathwayanalysis_wordcloud(pways, delimiter, excludeWords, ...)
}
\arguments{
\item{pways}{character vector. pathways to make a wordcloud with. or, more generally, a vector of "documents" with words separated by some delimiter.}

\item{delimiter}{a single character string. delimiter separating words in each of the elements of the vector \code{pways}}

\item{excludeWords}{words to exlude; for example, "GOBP"}

\item{...}{other arguments passed on to \code{wordcloud::wordcloud()}}
}
\value{
a wordcloud  as a \code{recordedplot} plot object.
}
\description{
This will make a wordcloud based on a character vector of pathway names.
If using with the rest of the pathway analysis pipeline, see the example below.
}
