% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{regressout}
\alias{regressout}
\title{Regress out a variable from a gene expression matrix with linear regression}
\usage{
regressout(gem, var_to_reg, scale = NULL, center = NULL)
}
\arguments{
\item{gem}{a matrix or data.frame; gene expression values with rows = genes aand columns = samples}

\item{var_to_reg}{a vector with length = ncol(gem) (ie a variable with some number for each sample)}

\item{scale}{T/F, default T, scale gem before lm}

\item{center}{T/F, default T, center gem before lm}
}
\description{
Remove the effect of a variable from a gene expression matrix. This is done with linear regression. For each gene, we take lm(gene expression ~ var_to_reg). We use the residuals from this linear model as the new expression values.
This was inspired by the Seurat::ScaleData function for single-cell analysis.
}
