% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival}
\alias{survival}
\title{A function for survival analysis.}
\usage{
survival(
  testvardf,
  clinvardf,
  vartypes,
  multivarnames,
  timevarname,
  statusvarname
)
}
\arguments{
\item{testvardf}{\itemize{
\item a data.frame of variables to test for association with survival, columns = variables, rows = observations. If you want to run a gene expression matrix, then genes need to be columns and samples need to be rows, ie you may need to transpose via the transpose function, t().
}}

\item{clinvardf}{\itemize{
\item a data.frame with clinical variables, including time and status columns, columns = variables, rows = observations.
}}

\item{vartypes}{\itemize{
\item a character vector of length = ncol(testvardf). Should say either "continuous" or "categorical", for each column of testvardf. If empty, will guess the vartype, continuous has >12 unique values.
}}

\item{multivarnames}{\itemize{
\item a character vector. The colnames of clinvardf to include in multivariable modelling, if desired. If empty, does not perform multivar modelling.
}}

\item{timevarname}{\itemize{
\item a string, the colname of the column in clinvardf with time to event information.
}}

\item{statusvarname}{\itemize{
\item a string, the colname of the column in clinvardf with event information. The event column should be a character vector and should have the word "Censored" (capital C); ie alive/dead should be "Censored" and "Dead", event should be "Censored" and "Event".
}}
}
\value{
A nested list object: highest level, categorical vs continuous; below that, lists containing plots, models, and data; plotlists will contain kaplan-meier plots; models will include cox regresion models; datalists contain the data DFs, which is important for reproducibility and model diagnostics including proportional hazards testing.
}
\description{
An easy wrapper around survival analysis as implemented in the packages survival and survminer.
}
