% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayanalysis.R
\name{unlist2}
\alias{unlist2}
\title{Unlist2 from AnnotationDBI package}
\usage{
unlist2(x, recursive = TRUE, use.names = TRUE, what.names = "inherited")
}
\arguments{
\item{what.names}{}
}
\description{
See \code{?AnnotationDbi::unlist2()} by Hervé Pagès
}
