% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gsea.results}
\alias{gsea.results}
\title{Get GSEA results}
\usage{
gsea.results(
  deg,
  pathways = NULL,
  nperm = NULL,
  weightmethod = NULL,
  onlypos = NULL
)
}
\arguments{
\item{deg}{dataframe with column names similar to the output of DESeq2 "results()" function}

\item{pathways}{list of character vectors containing gene names in same format as deg; each list element's name should be the name of the pathway}

\item{nperm}{how many permutations to run in FGSEA (may be deprecated). Default is 10000}

\item{weightmethod}{one of "pvalue" or "foldchange". how to weight the DEG results. Default is "Pvalue"}

\item{onlypos}{T/F. whether to use only positive log fold change genes. Default is False.}
}
\value{
a list with two elements; 1, a dataframe with FGSEA results; 2, a list of character vectors showing the leading edge genes for each pathway
}
\description{
Using FGSEA, DESeq2 results, and input pathways, compute GSEA results.
}
