% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEplots.R
\name{heatmapplot}
\alias{heatmapplot}
\title{Make an expression heatmap with hierarchical clustering}
\usage{
heatmapplot(
  expmatrix,
  genes,
  metadata,
  heatmaptitle,
  legendtitle,
  do.log2,
  do.scale,
  change_gene_label,
  gene_label_equivalency,
  clustering_distance_columns,
  ...
)
}
\arguments{
\item{expmatrix}{matrix or data.frame with rows = genes and columns = samples}

\item{genes}{character vector of gene names (rownames) in the expmatrix}

\item{metadata}{data.frame in DESeq2 "coldata" format. Needs a "Sample" column with sample IDs matching colnames of expmat, a "Condition" column with the associated conditon, and a "Color" column that has the sample colors}

\item{heatmaptitle}{string, name of heatmap on top}

\item{legendtitle}{string, what to call the values in the gene expression matrix. Will default to "Zscaled Normalized Counts"}

\item{do.log2}{T/F, whether to log2(x+1) transform gene expression values (rows) in the matrix}

\item{do.scale}{T/F, whether to scale (z-transform, (x - mean(x) / sd(x))) the gene expression values (rows) of the matrix}

\item{change_gene_label}{T/F, whether to change gene labels, defautl F}

\item{gene_label_equivalency}{data.frame, if change_gene_label is set to T, need to provide a data.frame with two columns, first column with current gene labels (rownames of res) and second column with gene labels you want to plot as labels. Useful for when you have res with IDs but want to show gene symbols.}

\item{clustering_distance_columns}{distance to use for ComplexHeatmap::Heatmap clustering of rows, default is pearson}

\item{...}{Additionaly arguments to pass to ComplexHeatmap::Heatmap}
}
\description{
assumes a normalized counts matrix, should work with TPM too. Currently, color depends on addding metadata. Later, will add support for manually defined colors.
}
