% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEplots.R
\name{heatmapplot}
\alias{heatmapplot}
\title{Make an expression heatmap with hierarchical clustering}
\usage{
heatmapplot(
  expmatrix,
  genes,
  metadata,
  heatmaptitle,
  legendtitle,
  do.log2,
  do.scale
)
}
\arguments{
\item{expmatrix}{matrix of data.frame with rows = genes and columns = samples}

\item{genes}{character vector of gene names (rownames) in the expmatrix}

\item{metadata}{data.frame in DESeq2 "coldata" format. Needs a "Sample" column with sample IDs matching colnames of expmat, a "Condition" column with the associated conditon, and a "Color" column that has the sample colors}

\item{heatmaptitle}{string, name of heatmap on top}

\item{legendtitle}{string, what to call the values in the gene expression matrix. Will default to "Zscaled Normalized Counts"}

\item{do.log2}{T/F, whether to log2(x+1) transform gene expression values (rows) in the matrix}

\item{do.scale}{T/F, whether to scale (z-transform, (x - mean(x) / sd(x))) the gene expression values (rows) of the matrix}
}
\value{

}
\description{
assumes a normalized counts matrix, should work with TPM too
}
\details{
For each gene, this will z-scale the normalized counts across samples. this is good for comparison
}
