% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival}
\alias{survival}
\title{A function for survival analysis.}
\usage{
survival(
  testvardf,
  clinvardf,
  vartypes,
  multivarnames,
  timevarname,
  statusvarname,
  outdir
)
}
\arguments{
\item{testvardf}{\itemize{
\item a data.frame of variables to test for association with survival, columns = variables, rows = observations. If you want to run a gene expression matrix, then genes need to be columns and samples need to be rows, ie you may need to transpose via the transpose function, t().
}}

\item{clinvardf}{\itemize{
\item a data.frame with clinical variables, including time and status columns, columns = variables, rows = observations.
}}

\item{vartypes}{\itemize{
\item a character vector of length = ncol(testvardf). Should say either "continuous" or "categorical", for each column of testvardf. If empty, will guess the vartype, defining continuous variables as those with >12 unique values.
}}

\item{multivarnames}{\itemize{
\item a character vector. The colnames of clinvardf to include in multivariable modelling, if desired. If empty, does not perform multivar modelling.
}}

\item{timevarname}{\itemize{
\item a string, the colname of the column in clinvardf with time to event information.
}}

\item{statusvarname}{\itemize{
\item a string, the colname of the column in clinvardf with event information. The event column should be a character vector and should have the word "Censored" (capital C); ie alive/dead should be "Censored" and "Dead", event should be "Censored" and "Event".
}}

\item{outdir}{\itemize{
\item a string, the path to write outputs to; defaults to './survival'
}}
}
\value{
Does not return anything; instead, will save survival analysis plots, Cox regression models, and frozen data to outdir.
}
\description{
An easy wrapper around survival analysis as implemented in the packages \code{survival} and \code{survminer}.
This package is meant to easily facilitate survival analysis for test variables (given in \code{testvardf}), using time-to-event and event-status variables (given in \code{clinvardf}).
It also supports multivariable modelling using other variables in \code{clinvardf}, which are called according to their column names as provided by \code{multivarnames}.
}
