% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEplots.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Make a VolcanoPlot}
\usage{
volcanoplot(
  results,
  metadata,
  contrast,
  condition1,
  condition2,
  colors,
  pval_thres,
  lfc_thres,
  lfclimit,
  pointsize,
  outliergene_pointsize,
  outliergene_shape,
  repeltextsize
)
}
\arguments{
\item{results}{dataframe, with column names similar to the output of DESeq2::results() function}

\item{metadata}{dataframe, for coloring samples using a consistently set color scheme. metadata is a dataframe similar to metadata used as input to DESEQ2; include a contrast column and colors coumn}

\item{contrast}{character string. name of contrast column in metadata. default will search for "Conditions"}

\item{condition1}{character string. will be used to find the corresponding contrast-color pairing for upregulated genes}

\item{condition2}{character string. will be used to find the corresponding contrast-color pairing for downregulated genes}

\item{colors}{character vector of length two with string names or codes of colors. if not using metadata-based coloring, will use this input or default to red/blue.}

\item{pval_thres}{numeric, alpha value cutoff for significant genes. default = 0.05}

\item{lfc_thres}{numeric, logFC cutoff for significant genes, will be symmetric on both sides. default = 0}

\item{lfclimit}{numeric, lofFC cutoff for outlier genes, will be symmetric on both sides. default = 15}

\item{pointsize}{numeric, point size of graph, passed to geom_point, default = 0.5}

\item{outliergene_pointsize}{numeric, point size of logFC outlier genes, default = 2}

\item{outliergene_shape}{numeric, code for shape symbol of logFC outlier genes, default = 6, which is a triangle}

\item{repeltextsize}{numeric, size of repel text gene labels, default is 2}
}
\value{
a ggplot2 object.
}
\description{
Using DESeq2 results, make a volcanoplot
}
\details{
color can be set with colors or passed using metadata. the latter is encouraged, in keeping with setting a color scheme at the start and keeping it consistent in all analysis.
}
