% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEplots.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Make a VolcanoPlot}
\usage{
volcanoplot(
  results,
  metadata,
  contrast,
  condition1,
  condition2,
  colors,
  pval_thres,
  lfc_thres,
  lfclimit,
  pointsize,
  outliergene_pointsize,
  outliergene_shape,
  repeltextsize
)
}
\arguments{
\item{results}{dataframe, with column names similar to the output of DESeq2::results() function}

\item{metadata}{dataframe, for coloring samples using a consistently set color scheme. metadata is a dataframe similar to coldata used in DESEQ2; include a contrast column and colors column}

\item{contrast}{character string. name of contrast column in metadata. default will search for "Conditions"}

\item{condition1}{character string. will be used to find the corresponding contrast-color pairing for upregulated genes}

\item{condition2}{character string. will be used to find the corresponding contrast-color pairing for downregulated genes}

\item{colors}{character vector of length two with string names or codes of colors. if not using metadata-based coloring, will use this input or default to red/blue.}

\item{pval_thres}{numeric, alpha value cutoff for significant genes. default = 0.05}

\item{lfc_thres}{numeric, logFC cutoff for significant genes, will be symmetric on both sides. default = 0}

\item{lfclimit}{numeric, lofFC cutoff for outlier genes, will be symmetric on both sides. default = 15}

\item{pointsize}{numeric, point size of graph, passed to geom_point, default = 0.5}

\item{outliergene_pointsize}{numeric, point size of logFC outlier genes, default = 2}

\item{outliergene_shape}{numeric, code for shape symbol of logFC outlier genes, default = 6, which is a triangle}

\item{repeltextsize}{numeric, size of repel text gene labels, default is 2}
}
\value{
a ggplot2 object.
}
\description{
Using DESeq2 results, make a volcanoplot
}
\details{
color can be set with colors or passed using metadata. the latter is encouraged, in keeping with setting a color scheme at the start and keeping it consistent in all analysis.
}
\examples{


With default colors:
volcanoplot(results)

With set colors:
volcanoplot(results, colors = c('Purple', 'Pink'))


To use metadata for coloration, we need a metadata dataframe
 that looks like this:

Samples   Condition    Color
Sample1   KO           Red
Sample2   KO           Red
Sample3   WT           Blue
Sample4   WT           Blue

contrast in this case can be input as contrast = "Condition"
metadata needs a column called Color, set it at the outset,
it will make all your plotting easier!
condition1 and condition2 should be set,
ie condition1 = "KO" and condition2 = "WT"

If we have all that, we can do this:
volcanoplot(results, metadata = metadata,
            condition1 = 'KO', condition2 = 'WT')


}
